# v2/api/openapi.yaml
openapi: 3.0.3
info:
  title: XRPL Excise Tax API
  description: |
    RESTful API for XRPL integration and excise tax calculations.
    
    This API provides endpoints to:
    - Query account balances and information  
    - Retrieve transaction details and history
    - Verify transaction status
    - Calculate excise taxes for alcoholic beverage production
    - Get current tax rates and product types
    
    All CLI functionality is accessible via these REST endpoints.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api-staging.example.com
    description: Staging server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /xrpl/accounts/{address}/info:
    get:
      tags:
        - Accounts
      summary: Get account information
      description: Retrieves detailed information for an XRPL account
      operationId: getAccountInfo
      parameters:
        - $ref: '#/components/parameters/AddressPath'
      responses:
        '200':
          description: Account information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /xrpl/accounts/{address}/transactions:
    get:
      tags:
        - Accounts
      summary: Get account transaction history
      description: Retrieves transaction history for an XRPL account with pagination
      operationId: getAccountTransactions
      parameters:
        - $ref: '#/components/parameters/AddressPath'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /xrpl/transactions/{hash}:
    get:
      tags:
        - Transactions
      summary: Get transaction details
      description: Retrieves detailed information about a specific transaction
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/HashPath'
      responses:
        '200':
          description: Transaction details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /tax/validate:
    post:
      tags:
        - Tax
      summary: Validate production data
      description: Validates production data without calculating taxes
      operationId: validateProductionData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionDataRequest'
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    AddressPath:
      name: address
      in: path
      required: true
      description: XRPL account address (starts with 'r')
      schema:
        type: string
        pattern: '^r[a-zA-Z0-9]{24,34}$'
        example: rPEPPER7kfTD9w2To4CQk6UCfuHM9c6GDY
    HashPath:
      name: hash
      in: path
      required: true
      description: Transaction hash (64-character hex string)
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{64}$'
        example: 3B4D6B3F5E2A1C8D9F0A7B6C5E4D3A2B1C0D9E8F7A6B5C4D3E2F1A0B9C8D7E6F
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Maximum number of transactions to return (1-100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    AccountInfoResponse:
      type: object
      required:
        - account
        - balance_xrp
        - balance_drops
        - sequence
      properties:
        account:
          type: string
        balance_xrp:
          type: string
        balance_drops:
          type: integer
        sequence:
          type: integer
          description: Account sequence number
        flags:
          type: integer
          description: Account flags
        owner_count:
          type: integer
          description: Number of objects owned
        previous_txn_id:
          type: string
          description: Hash of previous transaction
        previous_txn_lgr_seq:
          type: integer
          description: Ledger sequence of previous transaction
    TransactionsResponse:
      type: object
      required:
        - account
        - limit
        - count
        - transactions
      properties:
        account:
          type: string
        limit:
          type: integer
        count:
          type: integer
          description: Number of transactions returned
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    Transaction:
      type: object
      description: XRPL transaction object
      properties:
        hash:
          type: string
        ledger_index:
          type: integer
        date:
          type: integer
        tx:
          type: object
          description: Transaction details
    TransactionResponse:
      type: object
      required:
        - hash
        - transaction
      properties:
        hash:
          type: string
        transaction:
          $ref: '#/components/schemas/Transaction'
...
    ProductionData:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductionItem'
    ProductionItem:
      type: object
      required:
        - product_type
        - product_name
        - unit_type
        - quantity
      properties:
        product_type:
          type: string
          enum: [beer, wine, spirits, other]
        product_name:
          type: string
          example: IPA Batch #123
        unit_type:
          type: string
          enum: [gallon, barrel, case, liter]
        quantity:
          type: number
          format: float
          minimum: 0
          example: 100.5
        abv:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Alcohol by volume (percentage)
          example: 6.5
        notes:
          type: string
          example: Seasonal IPA
...
    ProductionDataRequest:
      type: object
      required:
        - production_data
      properties:
        production_data:
          $ref: '#/components/schemas/ProductionData'
    ValidationResponse:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: items[0].quantity
        message:
          type: string
          example: quantity must be greater than 0
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - type
            - title
            - status
          properties:
            type:
              type: string
              description: URI reference for error type
              example: https://api.example.com/errors/404
            title:
              type: string
              description: Short, human-readable error summary
              example: Not Found
            status:
              type: integer
              description: HTTP status code
              example: 404
            detail:
              type: string
              description: Human-readable explanation
              example: The requested resource could not be found
            instance:
              type: string
              description: URI reference for this specific error occurrence
              example: /xrpl/accounts/invalid_address/balance
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              type: https://api.example.com/errors/400
              title: Bad Request
              status: 400
              detail: invalid XRPL address format
              instance: /xrpl/accounts/invalid/balance
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              type: https://api.example.com/errors/404
              title: Not Found
              status: 404
              detail: The requested resource could not be found
              instance: /xrpl/accounts/rNotFound/balance
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              type: https://api.example.com/errors/500
              title: Internal Server Error
              status: 500
              detail: The server encountered an unexpected error
              instance: /xrpl/accounts/rSomeAccount/balance
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (for future use)
security: []